feat : detail/[id] -> auction/[id]로 이동 및 디자인 수정
